Move aggregate method.
